memperbaiki wawancara

+++ app/Http/Controllers/Admin/PertanyaanController.php
@@ -34,38 +34,48 @@ class PertanyaanController extends Controller
 
     public function save(Request $request)
     {
-        $validatedPertanyaan = $request->validate([
-            'pertanyaan' => ['required'],
-            'kategori' => ['required'],
-            'jawaban_benar' => ['required'],
-        ]);
-        $validatedJawaban = $request->validate([
-            'jawaban' => ['required'],
-        ]);
-
+        if($request->kategori == "TEORI"){
+            $validatedPertanyaan = $request->validate([
+                'pertanyaan' => ['required'],
+                'kategori' => ['required'],
+                'jawaban_benar' => ['required'],
+            ]);
+            $validatedJawaban = $request->validate([
+                'jawaban' => ['required'],
+            ]);
+        } else {
+            $validatedPertanyaan = $request->validate([
+                'pertanyaan' => ['required'],
+                'kategori' => ['required'],
+            ]);
+            $validatedPertanyaan['jawaban_benar'] = "Essay";
+        }
         $pertanyaan = Pertanyaan::create($validatedPertanyaan);
-        $validatedJawaban['pertanyaan_id'] = $pertanyaan->id;
 
-        $jawabanA = Jawaban::create([
-           'pertanyaan_id' => $validatedJawaban['pertanyaan_id'],
-           'pilihan' => "A",
-           'jawaban' => $validatedJawaban['jawaban']['A'], 
-        ]);
-        $jawabanB = Jawaban::create([
-           'pertanyaan_id' => $validatedJawaban['pertanyaan_id'],
-           'pilihan' => "B",
-           'jawaban' => $validatedJawaban['jawaban']['B'], 
-        ]);
-        $jawabanC = Jawaban::create([
-           'pertanyaan_id' => $validatedJawaban['pertanyaan_id'],
-           'pilihan' => "C",
-           'jawaban' => $validatedJawaban['jawaban']['C'], 
-        ]);
-        $jawabanD = Jawaban::create([
-           'pertanyaan_id' => $validatedJawaban['pertanyaan_id'],
-           'pilihan' => "D",
-           'jawaban' => $validatedJawaban['jawaban']['D'], 
-        ]);
+        if($request->kategori == "TEORI"){
+            $validatedJawaban['pertanyaan_id'] = $pertanyaan->id;
+    
+            $jawabanA = Jawaban::create([
+               'pertanyaan_id' => $validatedJawaban['pertanyaan_id'],
+               'pilihan' => "A",
+               'jawaban' => $validatedJawaban['jawaban']['A'], 
+            ]);
+            $jawabanB = Jawaban::create([
+               'pertanyaan_id' => $validatedJawaban['pertanyaan_id'],
+               'pilihan' => "B",
+               'jawaban' => $validatedJawaban['jawaban']['B'], 
+            ]);
+            $jawabanC = Jawaban::create([
+               'pertanyaan_id' => $validatedJawaban['pertanyaan_id'],
+               'pilihan' => "C",
+               'jawaban' => $validatedJawaban['jawaban']['C'], 
+            ]);
+            $jawabanD = Jawaban::create([
+               'pertanyaan_id' => $validatedJawaban['pertanyaan_id'],
+               'pilihan' => "D",
+               'jawaban' => $validatedJawaban['jawaban']['D'], 
+            ]);
+        }
         return redirect()->route('admin.pertanyaan.index')->with('success', 'Pertanyaan berhasil ditambahkan!');
     }
 
@@ -79,36 +89,47 @@ class PertanyaanController extends Controller
 
     public function update($id, Request $request)
     {
-        $validatedPertanyaan = $request->validate([
-            'pertanyaan' => ['required'],
-            'kategori' => ['required'],
-            'jawaban_benar' => ['required'],
-        ]);
-        $validatedJawaban = $request->validate([
-            'jawaban' => ['required'],
-        ]);
+        if($request->kategori == "TEORI"){
+            $validatedPertanyaan = $request->validate([
+                'pertanyaan' => ['required'],
+                'kategori' => ['required'],
+                'jawaban_benar' => ['required'],
+            ]);
+            $validatedJawaban = $request->validate([
+                'jawaban' => ['required'],
+            ]);
+        } else {
+            $validatedPertanyaan = $request->validate([
+                'pertanyaan' => ['required'],
+                'kategori' => ['required'],
+            ]);
+            $validatedPertanyaan['jawaban_benar'] = "Essay";
+        }
+
         $id = Crypt::decrypt($id);
         $pertanyaan = Pertanyaan::findOrFail($id);
         $pertanyaan->update($validatedPertanyaan);
+        if($request->kategori == "TEORI"){
+            $jawabanA = Jawaban::query()->where('pertanyaan_id', $id)->where('pilihan', '=', "A")->first();
+            $jawabanA->update([
+                'jawaban' => $validatedJawaban['jawaban']['A'],
+            ]);
+    
+            $jawabanB = Jawaban::query()->where('pertanyaan_id', $id)->where('pilihan', '=', "B")->first();
+            $jawabanB->update([
+                'jawaban' => $validatedJawaban['jawaban']['B'],
+            ]);
+            $jawabanC = Jawaban::query()->where('pertanyaan_id', $id)->where('pilihan', '=', "C")->first();
+            $jawabanC->update([
+                'jawaban' => $validatedJawaban['jawaban']['C'],
+            ]);
+            $jawabanD = Jawaban::query()->where('pertanyaan_id', $id)->where('pilihan', '=', "D")->first();
+            $jawabanD->update([
+                'jawaban' => $validatedJawaban['jawaban']['D'],
+            ]);
+            
+        }
 
-        $jawabanA = Jawaban::query()->where('pertanyaan_id', $id)->where('pilihan', '=', "A")->first();
-        $jawabanA->update([
-            'jawaban' => $validatedJawaban['jawaban']['A'],
-        ]);
-
-        $jawabanB = Jawaban::query()->where('pertanyaan_id', $id)->where('pilihan', '=', "B")->first();
-        $jawabanB->update([
-            'jawaban' => $validatedJawaban['jawaban']['B'],
-        ]);
-        $jawabanC = Jawaban::query()->where('pertanyaan_id', $id)->where('pilihan', '=', "C")->first();
-        $jawabanC->update([
-            'jawaban' => $validatedJawaban['jawaban']['C'],
-        ]);
-        $jawabanD = Jawaban::query()->where('pertanyaan_id', $id)->where('pilihan', '=', "D")->first();
-        $jawabanD->update([
-            'jawaban' => $validatedJawaban['jawaban']['D'],
-        ]);
-        
         return redirect()->route('admin.pertanyaan.index')->with('success', 'Data Pertanyaan berhasil diperbarui!');
     }
 
@@ -116,9 +137,11 @@ class PertanyaanController extends Controller
     {
         $id = Crypt::decrypt($id);
         $pertanyaan = Pertanyaan::findOrFail($id);
-        $jawabans = Jawaban::query()->where('pertanyaan_id', $id)->get();
-        foreach ($jawabans as $jw){
-            $jw->delete();
+        if($pertanyaan->kategori == "TEORI"){
+            $jawabans = Jawaban::query()->where('pertanyaan_id', $id)->get();
+            foreach ($jawabans as $jw){
+                $jw->delete();
+            }
         }
         $pertanyaan->delete();
         return redirect()->route('admin.pertanyaan.index')->with('success', 'Data pertanyaan berhasil dihapus!');
